Popup Notepad in the view information layout
*Moved the Popup Notepad to view information layout
*Deleted Notepad in the Edit information
*Save and Cancel Functionalities done
