+++ java/citaapp/citapprenticeshipapp/ViewInformation.java
@@ -5,9 +5,13 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.Gravity;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.PopupWindow;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -16,14 +20,22 @@ public class ViewInformation extends AppCompatActivity
     SharedPreferences preferenceSettings;
     SharedPreferences.Editor preferenceEditor;
 
-    EditText txtUsi, txtCitNum, txtAnpName, txtAnpPhone, txtAnpEmail, txtDeptName, txtDeptPhone, txtDeptEmail, txtLlnDate, txtClassDate;
+    EditText txtUsi, txtCitNum, txtAnpName, txtAnpPhone, txtAnpEmail, txtDeptName, txtDeptPhone, txtDeptEmail, txtLlnDate, txtClassDate, txtEditNote;
 
-    TextView RPLlbl, TPClbl, txtViewNote;
+    TextView RPLlbl, TPClbl;
 
-    Button editBtn;
+    Button editBtn, BackBtn, btnSaveNote, btnCancelNote;
+    LayoutInflater layoutInflater;
+    PopupWindow popupNotepad;
 
-    Button BackBtn;
+    private void saveToPrefs(String key, String value)
+    {
+        preferenceSettings = getSharedPreferences("myInfo",Context.MODE_PRIVATE);
+        preferenceEditor = preferenceSettings.edit();
 
+        preferenceEditor.putString(key, value);
+        preferenceEditor.apply();
+    }
     void loadPrefs()
     {
         preferenceSettings = getSharedPreferences("myInfo", Context.MODE_PRIVATE);
@@ -41,7 +53,7 @@ public class ViewInformation extends AppCompatActivity
             txtDeptEmail.setText(preferenceSettings.getString("Department Email", "Not Found"));
             txtLlnDate.setText(preferenceSettings.getString("LLN Date", "Not Found"));
             txtClassDate.setText(preferenceSettings.getString("Class Start Date", "Not Found"));
-            txtViewNote.setText(preferenceSettings.getString("Notepad", "Notepad"));
+            txtEditNote.setText(preferenceSettings.getString("Notepad", "Notepad"));
             RPLlbl.setText(preferenceSettings.getString("RPL", "No"));
             TPClbl.setText(preferenceSettings.getString("Training Plan", "No"));
 
@@ -70,13 +82,59 @@ public class ViewInformation extends AppCompatActivity
         txtDeptEmail = (EditText) findViewById(R.id.txtDeptEmail);
         txtLlnDate = (EditText) findViewById(R.id.txtLlnDate);
         txtClassDate = (EditText) findViewById(R.id.txtClassDate);
-        txtViewNote = (TextView) findViewById(R.id.txtViewNote);
+        txtEditNote = (EditText) findViewById(R.id.txtEditNote);
         RPLlbl = (TextView) findViewById(R.id.lblRPL);
         TPClbl = (TextView) findViewById(R.id.lblTPC);
 
 
 
         editBtn = (Button) findViewById(R.id.btnEdit);
+
+        //triggers when the user clicks on the notepad
+        txtEditNote.setOnClickListener( new View.OnClickListener(){
+
+            @Override
+            public void onClick(View v) {
+                //creates a popup Window
+                layoutInflater = (LayoutInflater)getApplicationContext().getSystemService(LAYOUT_INFLATER_SERVICE);
+                ViewGroup container = (ViewGroup)layoutInflater.inflate(R.layout.popup_notepad,null);
+                popupNotepad = new PopupWindow(container, ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.FILL_PARENT,true);
+
+                popupNotepad.showAtLocation(v, Gravity.CENTER,0,0);
+                //get contents from edit text
+
+                final EditText notepadEditText;
+                notepadEditText = ((EditText)popupNotepad.getContentView().findViewById(R.id.editTextNotepad));
+                notepadEditText.setText(txtEditNote.getText());
+
+                btnCancelNote = (Button)popupNotepad.getContentView().findViewById(R.id.btnCancelNote);
+                btnSaveNote = (Button)popupNotepad.getContentView().findViewById(R.id.btnSaveNote);
+                btnCancelNote.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view)
+                    {
+
+                        popupNotepad.dismiss();
+                    }
+                });
+
+                btnSaveNote.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view)
+                    {
+                        try {
+                            saveToPrefs("Notepad", notepadEditText.getText().toString());
+                            txtEditNote.setText(notepadEditText.getText());
+                            popupNotepad.dismiss();
+                        }
+                        catch(Exception e)
+                        {
+                            Toast.makeText(getApplicationContext(), e.getMessage(), Toast.LENGTH_LONG).show();
+                        }
+                    }
+                });
+            }
+        });
 
         editBtn.setOnClickListener(new View.OnClickListener() {
             @Override
