+++ res/layout/activity_edit_information.xml
@@ -450,24 +450,7 @@
                 android:maxLines="1"
                 /><!-- need to set this focusableInTouchMode=false so that click event to bring up date picker works properly.-->
 
-            <EditText
-                android:text="NotePad"
-                android:layout_width="match_parent"
-                android:layout_height="200dp"
-                android:scrollbars="vertical"
-                android:paddingLeft="10dp"
-                android:paddingRight="10dp"
-                android:paddingTop="20dp"
-                android:layout_marginTop="10dp"
-                android:background="@drawable/notelist"
-                android:focusable="false"
-                android:id="@+id/txtEditNote"
-                android:maxLines="10"
-                android:maxLength="300"
-                android:textSize="15sp"
-                android:gravity="top"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true" />
+
 
         </LinearLayout>
     </RelativeLayout>
