+++ res/layout/activity_view_information.xml
@@ -431,20 +431,24 @@
                 android:layout_marginTop="10dp"
                 android:enabled="false"
                 android:focusable="false" />
-            <TextView
+            <EditText
                 android:text="NotePad"
                 android:layout_width="match_parent"
                 android:layout_height="200dp"
-                android:layout_marginTop="10dp"
-                android:paddingTop="20dp"
+                android:scrollbars="vertical"
                 android:paddingLeft="10dp"
                 android:paddingRight="10dp"
+                android:paddingTop="20dp"
+                android:layout_marginTop="10dp"
                 android:background="@drawable/notelist"
-                android:id="@+id/txtViewNote"
-                android:textSize="14sp"
+                android:focusable="false"
+                android:id="@+id/txtEditNote"
+                android:maxLines="10"
+                android:maxLength="300"
+                android:textSize="15sp"
+                android:gravity="top"
                 android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true"
-                />
+                android:layout_alignParentStart="true" />
         </LinearLayout>
